move helper functions to utils file for re-usability. Update mock youtube data to remove gifs
