+++ src/Utils/mock-youtube-date.js
@@ -17,7 +17,7 @@ export default [
     published_at: '2018-07-08T18:28:49.514Z',
     description: 'Lorem ipsum dolor amet ethical tbh sint master cleanse, in sed occupy gluten-free readymade PBR&B ex non officia vice pour-over. Kogi everyday carry tumblr, retro squid literally cardigan selvage echo park occupy neutra hoodie. Offal organic brooklyn DIY, ramps meditation in chillwave. Artisan before they sold out distillery four loko portland. Pug eu cornhole labore viral, sint post-ironic dolor art party hot chicken leggings gastropub readymade velit.',
     title: 'How to Fly Anywher for free',
-    thumbnail_url: 'https://i.ytimg.com/an_webp/9WyZWvpZiB4/mqdefault_6s.webp?du=3000&sqp=CNbujekF&rs=AOn4CLC4ZCfVCtVX_OKQ3GBccx_gFFdSUw'
+    thumbnail_url: 'https://i.ytimg.com/vi/8YbZuaBP9B8/maxresdefault.jpg'
   },
   {
     video_length: 'PT15M4S',
@@ -27,7 +27,7 @@ export default [
     published_at: '2019-03-08T18:28:49.514Z',
     description: 'Food truck actually glossier vexillologist sriracha meditation gochujang. Drinking vinegar offal before they sold out DIY affogato gentrify. Kogi skateboard tote bag four loko ugh try-hard godard cronut umami. Thundercats flannel kombucha chartreuse neutra 90\'s prism church-key tote bag vinyl typewriter activated charcoal cronut authentic coloring book. Health goth paleo forage man braid, yr pok pok listicle.',
     title: 'Don\'t Make Your next Kite Without Watching This',
-    thumbnail_url: 'https://i.ytimg.com/an_webp/Ba9M3lx05h4/mqdefault_6s.webp?du=3000&sqp=CNqHjukF&rs=AOn4CLDwbdFGE_Vmh4e1aDerO1TBP6WZkw'
+    thumbnail_url: 'https://i.ytimg.com/vi/ReE87Wl6c4I/maxresdefault.jpg'
   },
   {
     video_length: 'PT36M19S',
@@ -36,6 +36,7 @@ export default [
     youtube_display_name: 'Christopher Lau',
     published_at: '2017-01-08T18:28:49.514Z',
     description: '2018 was the year I left my 9-5 to pursue the passion of travel and filmmaking. It was the most incredible year of my life as I met some amazing friends including the love of my life. I started off in Indonesia, Australia, Cambodia, Thailand, Vietnam, Laos, & The Philippines. Now I have made Thailand my home base and I couldn\'t be happier. This isn\'t a Sam Kolder type transition everywhere video, but a See Lau Travel type video ;) Enjoy! #SeeLauTravel',
-    thumbnail_url: 'https://i.ytimg.com/an_webp/3q6_H5zHzQI/mqdefault_6s.webp?du=3000&sqp=CK6LjukF&rs=AOn4CLCGByyuAjEYYYE3KqEk3eWi_qoLtg'
+    title: '12 Amazing Dinner Party Ideas to Wow Your Friends',
+    thumbnail_url: 'https://i.ytimg.com/vi/4Yptevp7eeM/maxresdefault.jpg'
   }
 ]
